+++ Model/Base/SocolissimoDeliveryModeQuery.php
@@ -41,6 +41,14 @@ use SoColissimo\Model\Map\SocolissimoDeliveryModeTableMap;
  * @method     ChildSocolissimoDeliveryModeQuery rightJoinSocolissimoPrice($relationAlias = null) Adds a RIGHT JOIN clause to the query using the SocolissimoPrice relation
  * @method     ChildSocolissimoDeliveryModeQuery innerJoinSocolissimoPrice($relationAlias = null) Adds a INNER JOIN clause to the query using the SocolissimoPrice relation
  *
+ * @method     ChildSocolissimoDeliveryModeQuery leftJoinSocolissimoAreaFreeshippingDom($relationAlias = null) Adds a LEFT JOIN clause to the query using the SocolissimoAreaFreeshippingDom relation
+ * @method     ChildSocolissimoDeliveryModeQuery rightJoinSocolissimoAreaFreeshippingDom($relationAlias = null) Adds a RIGHT JOIN clause to the query using the SocolissimoAreaFreeshippingDom relation
+ * @method     ChildSocolissimoDeliveryModeQuery innerJoinSocolissimoAreaFreeshippingDom($relationAlias = null) Adds a INNER JOIN clause to the query using the SocolissimoAreaFreeshippingDom relation
+ *
+ * @method     ChildSocolissimoDeliveryModeQuery leftJoinSocolissimoAreaFreeshippingPr($relationAlias = null) Adds a LEFT JOIN clause to the query using the SocolissimoAreaFreeshippingPr relation
+ * @method     ChildSocolissimoDeliveryModeQuery rightJoinSocolissimoAreaFreeshippingPr($relationAlias = null) Adds a RIGHT JOIN clause to the query using the SocolissimoAreaFreeshippingPr relation
+ * @method     ChildSocolissimoDeliveryModeQuery innerJoinSocolissimoAreaFreeshippingPr($relationAlias = null) Adds a INNER JOIN clause to the query using the SocolissimoAreaFreeshippingPr relation
+ *
  * @method     ChildSocolissimoDeliveryMode findOne(ConnectionInterface $con = null) Return the first ChildSocolissimoDeliveryMode matching the query
  * @method     ChildSocolissimoDeliveryMode findOneOrCreate(ConnectionInterface $con = null) Return the first ChildSocolissimoDeliveryMode matching the query, or a new ChildSocolissimoDeliveryMode object populated from the query conditions when no match is found
  *
@@ -470,6 +478,152 @@ abstract class SocolissimoDeliveryModeQuery extends ModelCriteria
         return $this
             ->joinSocolissimoPrice($relationAlias, $joinType)
             ->useQuery($relationAlias ? $relationAlias : 'SocolissimoPrice', '\SoColissimo\Model\SocolissimoPriceQuery');
+    }
+
+    /**
+     * Filter the query by a related \SoColissimo\Model\SocolissimoAreaFreeshippingDom object
+     *
+     * @param \SoColissimo\Model\SocolissimoAreaFreeshippingDom|ObjectCollection $socolissimoAreaFreeshippingDom  the related object to use as filter
+     * @param string $comparison Operator to use for the column comparison, defaults to Criteria::EQUAL
+     *
+     * @return ChildSocolissimoDeliveryModeQuery The current query, for fluid interface
+     */
+    public function filterBySocolissimoAreaFreeshippingDom($socolissimoAreaFreeshippingDom, $comparison = null)
+    {
+        if ($socolissimoAreaFreeshippingDom instanceof \SoColissimo\Model\SocolissimoAreaFreeshippingDom) {
+            return $this
+                ->addUsingAlias(SocolissimoDeliveryModeTableMap::ID, $socolissimoAreaFreeshippingDom->getDeliveryModeId(), $comparison);
+        } elseif ($socolissimoAreaFreeshippingDom instanceof ObjectCollection) {
+            return $this
+                ->useSocolissimoAreaFreeshippingDomQuery()
+                ->filterByPrimaryKeys($socolissimoAreaFreeshippingDom->getPrimaryKeys())
+                ->endUse();
+        } else {
+            throw new PropelException('filterBySocolissimoAreaFreeshippingDom() only accepts arguments of type \SoColissimo\Model\SocolissimoAreaFreeshippingDom or Collection');
+        }
+    }
+
+    /**
+     * Adds a JOIN clause to the query using the SocolissimoAreaFreeshippingDom relation
+     *
+     * @param     string $relationAlias optional alias for the relation
+     * @param     string $joinType Accepted values are null, 'left join', 'right join', 'inner join'
+     *
+     * @return ChildSocolissimoDeliveryModeQuery The current query, for fluid interface
+     */
+    public function joinSocolissimoAreaFreeshippingDom($relationAlias = null, $joinType = Criteria::INNER_JOIN)
+    {
+        $tableMap = $this->getTableMap();
+        $relationMap = $tableMap->getRelation('SocolissimoAreaFreeshippingDom');
+
+        // create a ModelJoin object for this join
+        $join = new ModelJoin();
+        $join->setJoinType($joinType);
+        $join->setRelationMap($relationMap, $this->useAliasInSQL ? $this->getModelAlias() : null, $relationAlias);
+        if ($previousJoin = $this->getPreviousJoin()) {
+            $join->setPreviousJoin($previousJoin);
+        }
+
+        // add the ModelJoin to the current object
+        if ($relationAlias) {
+            $this->addAlias($relationAlias, $relationMap->getRightTable()->getName());
+            $this->addJoinObject($join, $relationAlias);
+        } else {
+            $this->addJoinObject($join, 'SocolissimoAreaFreeshippingDom');
+        }
+
+        return $this;
+    }
+
+    /**
+     * Use the SocolissimoAreaFreeshippingDom relation SocolissimoAreaFreeshippingDom object
+     *
+     * @see useQuery()
+     *
+     * @param     string $relationAlias optional alias for the relation,
+     *                                   to be used as main alias in the secondary query
+     * @param     string $joinType Accepted values are null, 'left join', 'right join', 'inner join'
+     *
+     * @return   \SoColissimo\Model\SocolissimoAreaFreeshippingDomQuery A secondary query class using the current class as primary query
+     */
+    public function useSocolissimoAreaFreeshippingDomQuery($relationAlias = null, $joinType = Criteria::INNER_JOIN)
+    {
+        return $this
+            ->joinSocolissimoAreaFreeshippingDom($relationAlias, $joinType)
+            ->useQuery($relationAlias ? $relationAlias : 'SocolissimoAreaFreeshippingDom', '\SoColissimo\Model\SocolissimoAreaFreeshippingDomQuery');
+    }
+
+    /**
+     * Filter the query by a related \SoColissimo\Model\SocolissimoAreaFreeshippingPr object
+     *
+     * @param \SoColissimo\Model\SocolissimoAreaFreeshippingPr|ObjectCollection $socolissimoAreaFreeshippingPr  the related object to use as filter
+     * @param string $comparison Operator to use for the column comparison, defaults to Criteria::EQUAL
+     *
+     * @return ChildSocolissimoDeliveryModeQuery The current query, for fluid interface
+     */
+    public function filterBySocolissimoAreaFreeshippingPr($socolissimoAreaFreeshippingPr, $comparison = null)
+    {
+        if ($socolissimoAreaFreeshippingPr instanceof \SoColissimo\Model\SocolissimoAreaFreeshippingPr) {
+            return $this
+                ->addUsingAlias(SocolissimoDeliveryModeTableMap::ID, $socolissimoAreaFreeshippingPr->getDeliveryModeId(), $comparison);
+        } elseif ($socolissimoAreaFreeshippingPr instanceof ObjectCollection) {
+            return $this
+                ->useSocolissimoAreaFreeshippingPrQuery()
+                ->filterByPrimaryKeys($socolissimoAreaFreeshippingPr->getPrimaryKeys())
+                ->endUse();
+        } else {
+            throw new PropelException('filterBySocolissimoAreaFreeshippingPr() only accepts arguments of type \SoColissimo\Model\SocolissimoAreaFreeshippingPr or Collection');
+        }
+    }
+
+    /**
+     * Adds a JOIN clause to the query using the SocolissimoAreaFreeshippingPr relation
+     *
+     * @param     string $relationAlias optional alias for the relation
+     * @param     string $joinType Accepted values are null, 'left join', 'right join', 'inner join'
+     *
+     * @return ChildSocolissimoDeliveryModeQuery The current query, for fluid interface
+     */
+    public function joinSocolissimoAreaFreeshippingPr($relationAlias = null, $joinType = Criteria::INNER_JOIN)
+    {
+        $tableMap = $this->getTableMap();
+        $relationMap = $tableMap->getRelation('SocolissimoAreaFreeshippingPr');
+
+        // create a ModelJoin object for this join
+        $join = new ModelJoin();
+        $join->setJoinType($joinType);
+        $join->setRelationMap($relationMap, $this->useAliasInSQL ? $this->getModelAlias() : null, $relationAlias);
+        if ($previousJoin = $this->getPreviousJoin()) {
+            $join->setPreviousJoin($previousJoin);
+        }
+
+        // add the ModelJoin to the current object
+        if ($relationAlias) {
+            $this->addAlias($relationAlias, $relationMap->getRightTable()->getName());
+            $this->addJoinObject($join, $relationAlias);
+        } else {
+            $this->addJoinObject($join, 'SocolissimoAreaFreeshippingPr');
+        }
+
+        return $this;
+    }
+
+    /**
+     * Use the SocolissimoAreaFreeshippingPr relation SocolissimoAreaFreeshippingPr object
+     *
+     * @see useQuery()
+     *
+     * @param     string $relationAlias optional alias for the relation,
+     *                                   to be used as main alias in the secondary query
+     * @param     string $joinType Accepted values are null, 'left join', 'right join', 'inner join'
+     *
+     * @return   \SoColissimo\Model\SocolissimoAreaFreeshippingPrQuery A secondary query class using the current class as primary query
+     */
+    public function useSocolissimoAreaFreeshippingPrQuery($relationAlias = null, $joinType = Criteria::INNER_JOIN)
+    {
+        return $this
+            ->joinSocolissimoAreaFreeshippingPr($relationAlias, $joinType)
+            ->useQuery($relationAlias ? $relationAlias : 'SocolissimoAreaFreeshippingPr', '\SoColissimo\Model\SocolissimoAreaFreeshippingPrQuery');
     }
 
     /**