+++ Model/SocolissimoAreaFreeshippingDom.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SoColissimo\Model;
+
+use SoColissimo\Model\Base\SocolissimoAreaFreeshippingDom as BaseSocolissimoAreaFreeshippingDom;
+
+class SocolissimoAreaFreeshippingDom extends BaseSocolissimoAreaFreeshippingDom
+{
+
+}