+++ Model/SocolissimoAreaFreeshippingDomQuery.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace SoColissimo\Model;
+
+use SoColissimo\Model\Base\SocolissimoAreaFreeshippingDomQuery as BaseSocolissimoAreaFreeshippingDomQuery;
+
+
+/**
+ * Skeleton subclass for performing query and update operations on the 'socolissimo_area_freeshipping_dom' table.
+ *
+ *
+ *
+ * You should add additional methods to this class to meet the
+ * application requirements.  This class will only be generated as
+ * long as it does not already exist in the output directory.
+ *
+ */
+class SocolissimoAreaFreeshippingDomQuery extends BaseSocolissimoAreaFreeshippingDomQuery
+{
+
+} // SocolissimoAreaFreeshippingDomQuery