+++ Model/SocolissimoAreaFreeshippingPr.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SoColissimo\Model;
+
+use SoColissimo\Model\Base\SocolissimoAreaFreeshippingPr as BaseSocolissimoAreaFreeshippingPr;
+
+class SocolissimoAreaFreeshippingPr extends BaseSocolissimoAreaFreeshippingPr
+{
+
+}